third_party/WebKit/LayoutTests/editing/selection/to_string/toString-1.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div><div id="sample" contenteditable>aaaa<object></object><div>bbbb</div>
</div></div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    selection.setBaseAndExtent(sample, 1, sample, 3);
    assert_equals(selection.toString(), "\nbbbb", "toString");
});
</script>
